Create wrap.py
Python source for the wrap function.  Grabs monitor details from Windows, builds a list of monitors, including size information, so that borders are known.  Tracks mouse events and acts on mouse movement, detecting when the edge of a boundary is hit and moving over to the opposing boundary, where movement continues.